Put things in the right path

+++ seraphs/generate.py
@@ -4,18 +4,19 @@
 # This work is in the public domain http://creativecommons.org/publicdomain/zero/1.0/
 # Libraries used by this work may be commercial or have other copyright restrictions.
 
-import os.path
+import os
 import flickr
 import random
 import shutil
 
 from jinja2 import Environment, PackageLoader
-from seraphs import BUILD_DIR
+import json
+from seraphs import BUILD_DIR, THIS_DIR
+
 env = Environment(loader=PackageLoader('seraphs', 'templates'))
 
 # The sections of the books that will be thematically related, based on the Voynich sections
 BOOK_SECTIONS = ('herb', 'astronomy', 'biology', 'astrology', 'medicine', 'alchemy',)
-THIS_DIR = os.path.dirname(os.path.realpath(__file__))
 
 def fill_template_page(section, images, words):
     template = env.get_template('1.html')
@@ -24,7 +25,8 @@ def fill_template_page(section, images, words):
 if __name__ == '__main__':
 
     # Delete previous runs and re-create build dir
-    shutil.rmtree(BUILD_DIR)
+    if os.path.exists(BUILD_DIR):
+        shutil.rmtree(BUILD_DIR)
     os.makedirs(BUILD_DIR)
         
     # Delete
@@ -34,4 +36,5 @@ if __name__ == '__main__':
     for section in BOOK_SECTIONS[0:1]:
         images = flickr.flickr_search(section)
         rendered_template = fill_template_page(section, images, words)
-        print rendered_template        
+        out = open(os.path.join(BUILD_DIR, 'book.html'), 'w')
+        out.write(rendered_template)
